% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geral.R
\name{consertar_documentos}
\alias{consertar_documentos}
\title{Conserta documentos}
\usage{
consertar_documentos(x)
}
\arguments{
\item{x}{O documento ou vetor de documentos a serem consertados.}
}
\value{
O documento ou vetor de documentos corrigidos.
}
\description{
Esta função recebe um documento ou vetor de documentos e conserta seu formato,
completando com zeros à esquerda se necessário.
}
