% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_operations.R
\name{mean}
\alias{mean}
\title{Title}
\usage{
mean(x, na.rm = T, ...)

mean(x, na.rm = T, ...)
}
\arguments{
\item{x}{numeric or complex or logical vectors.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\description{
Title

Title
}
