% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geral.R
\name{remover_colunas_vazias}
\alias{remover_colunas_vazias}
\title{Remove colunas com todos os valores vazios ou repetidos}
\usage{
remover_colunas_vazias(df)
}
\arguments{
\item{df}{O data frame a ser processado.}
}
\value{
Um novo data frame sem as colunas mencionadas.
}
\description{
Esta função remove as colunas de um data frame que possuem todos os valores vazios
ou contêm o mesmo valor repetido em todas as linhas.
}
