% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_operations.R
\name{sum}
\alias{sum}
\title{Title}
\usage{
sum(..., na.rm = T)
}
\arguments{
\item{...}{numeric or complex or logical vectors.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\description{
Title
}
