% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratar_bdc_family.R
\name{trata_bdc_processo_socios}
\alias{trata_bdc_processo_socios}
\title{Trata Processo de Sócios}
\usage{
trata_bdc_processo_socios(initial_path, dest_folder)
}
\arguments{
\item{initial_path}{O caminho para o arquivo inicial.}

\item{dest_folder}{O diretório de destino para salvar o arquivo tratado.}
}
\value{
Nenhum valor é retornado explicitamente, o arquivo tratado é salvo no diretório de destino.
}
\description{
Esta função realiza o tratamento de dados de processos de sócios.
}
