% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratar_bdc_family.R
\name{unnest_all}
\alias{unnest_all}
\title{Função unnest_all}
\usage{
unnest_all(df, recursively = TRUE, ...)
}
\arguments{
\item{df}{O data frame a ser expandido.}

\item{recursively}{Um valor lógico indicando se a função deve ser aplicada recursivamente.}

\item{...}{Outros argumentos a serem passados para a função \code{unnest()}.}
}
\value{
O data frame expandido.
}
\description{
Esta função expande todas as colunas que contêm listas ou data frames em um data frame.
}
