% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_recursivo.R
\name{unnest_recursivo}
\alias{unnest_recursivo}
\title{Aplica unnest de forma recursiva em colunas de data frame ou lista}
\usage{
unnest_recursivo(df)
}
\arguments{
\item{df}{O data frame ou lista a ser processado.}
}
\value{
O data frame resultante após aplicar unnest de forma recursiva.
}
\description{
Esta função aplica a função unnest de forma recursiva em colunas de um data frame ou lista,
expandindo as colunas que são data frames ou listas até que não haja mais aninhamentos.
}
\examples{
df <- data.frame(
  col1 = list(1, list(2, 3)),
  col2 = list(data.frame(a = 1:2, b = 3:4), data.frame(c = 5:6))
)
unnest_recursivo(df)

}
